Made `make` text clearer for unfamiliar people.
I initially just fixed some typos and grammar, but then also decided to improve the text for the sake of anyone less familiar with what the text is talking about.

Also, even I myself almost accidentally forgot the download button at the top, so I added a comment about that too so that Windows user's aren't accidentally scared off.

Support for `make` on Windows is often very poor and relying upon it in a build process will signal to people that it the package may not be usable for contributions on some Windows systems. So, it's thus very important to make people aware that the pre-built Windows binaries exist probably.