Unclear "Example" header, so made it more specific
The header "Example" made it slightly unclear what it was an example of. For example, the reader may have wondered if it was supposed to be another example related to `rl.ref` but then wouldn't realize it wasn't intended that way until after finishing reading the example code.

So, I changed the title here to clarify that it is a more general example for just testing the system as a whole.