Fixed some grammar and formatting for the compatibility section
I'm personally confused by the mention of the "raylib 3.5 API" when the raylib bindings page claims this raylib binding library actually supports raylib 4.5.

Perhaps I don't understand though. Is a part of the library under a different version? Is part in 3.5 while another part is in 4.5?

I also added links to what I thought is the most likely things you are referring, but I'm not 100% sure on those either.